add relations and change neccessary columns to Byte (tinyint in sql)

+++ src/main/java/sk/f1api/f1api/entity/SessionType.java
@@ -11,7 +11,7 @@ public class SessionType {
 
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Byte id;
 
     @Column(nullable = false, unique = true, length = 2, columnDefinition="CHAR")
     private String abbreviation;
